% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstrap.R
\name{upstrap}
\alias{upstrap}
\title{Upstrap Resampling}
\usage{
upstrap(data, statistic, R = 10000, n.grid.multiplier = 21,
  range_multiplier = c(1, 5), ...)
}
\arguments{
\item{data}{The data as a matrix or data frame.
If it is a matrix or data frame then each row
is considered as one multivariate observation.}

\item{statistic}{A function which when applied to data returns a vector
containing the statistic(s) of interest.}

\item{R}{The number of upstrap replicates for each point on the grid of
range multipliers. Usually this will be a single positive integer.}

\item{n.grid.multiplier}{number of grid points for the
multiplier of the sample size.}

\item{range_multiplier}{minimum and maximum multiplier for the sample size}

\item{...}{additional arguments to pass to \code{statistic}.}
}
\value{
An object
}
\description{
Upstrap Resampling
}
\examples{
fname = system.file("extdata", "shhs1.txt",
package = "upstrap")
# read in the data
data = read.table(file = fname,
                 header = TRUE, na.strings="NA")
data$MtS_SA = data$rdi4p>=15
statistic = function(data) {
fit<-glm(MtS_SA ~ gender + age_s1 + bmi_s1 +
HTNDerv_s1 + age_s1*HTNDerv_s1, family="binomial", data=data)
# could use broom
#  tid = broom::tidy(fit)
#  pval = tid$p.value[ tid$term == "HTNDerv_s1"]
smod = coef(summary(fit))
pval = smod["HTNDerv_s1", "Pr(>|z|)"]
pval < 0.05
}
res = upstrap(data, statistic = statistic, R = 100, n.grid.multiplier = 5,
range_multiplier = c(1, 3))
res = colMeans(res)
}
